requirements.txt
streamlit==1.43.2
fastapi==0.115.11
pydantic==2.10.6
uvicorn==0.30.0
mysql-connector-python==9.2.0
pandas==2.2.3
requests==2.32.3
pytest==8.3.5